+++ src/main/java/dev/ninesliced/commands/BetterMapMenuCommand.java
@@ -17,8 +17,8 @@ public class BetterMapMenuCommand extends AbstractPlayerCommand {
     }
 
     @Override
-    protected String generatePermissionNode() {
-        return "";
+    protected boolean canGeneratePermission() {
+        return false;
     }
 
     @Override