+++ src/main/java/dev/ninesliced/commands/BetterMapWaypointDeleteCommand.java
@@ -23,8 +23,8 @@ public class BetterMapWaypointDeleteCommand extends AbstractPlayerCommand {
     private final RequiredArg<String> targetArg = this.withRequiredArg("target", "Waypoint name or marker id", ArgTypes.STRING);
 
     @Override
-    protected String generatePermissionNode() {
-        return "";
+    protected boolean canGeneratePermission() {
+        return false;
     }
 
     public BetterMapWaypointDeleteCommand() {