+++ src/main/java/dev/ninesliced/commands/config/AutoSaveIntervalCommand.java
@@ -28,7 +28,7 @@ public class AutoSaveIntervalCommand extends AbstractCommand {
 
     @Override
     protected String generatePermissionNode() {
-        return "config.autosave";
+        return "autosave";
     }
 
     @Nullable